Consolidate initial database migration
Database models and migrations get changed rather often during the
early development and data design. Put all migrations into a single file
to make handling and work with them easier.

Also, change to relationship between `Question` and `Answer` from
Many-to-Many to One-To-Many. While questions sometimes have identical
answers, the implementation of other features such as answer comments or
statistics are less complex if an answer belongs to only a single
question.
For the few cases where an answer is identical for multile questions, we
accept the redundancy in favor of a simpler implementation.