SessionImageView: improve fullscreen image view
Currently, the image is shown as a card and not actually "fluid". It
should either have its max size or match the size of the view port.

Also, make sure to properly set and unset the hotkeys scope in order to
not interfere with the question navigation.

Finally, use a "key'ed" `{#each ...}` block for images to make Svelte
regenerate the image element when the user switches to the next
questions. Otherwise the next question will show the image of the
previous question.