SessionAnswerView: give excluded answer options red background
Users can mark an answer option as "probably wrong" by clicking a
"close" button on an answer. Currently, this does strike through the
text of the answer, which is not a very obvious marker. Instead, give
the answer a red background and keep the text readable at the same time.