NewSessionView: use `list-group`s for subjects/modules
It's easier to grasp the available subjects and modules when they are
shown as a list. Also, the current code with `select` items has a bug in
the `bind:value=...` logic and always shows the first module of a
subject.

We should remember the last selected subject and module for the user.
This remains to be done.