SessionProgressBar: refactor interface
Refactor the interface of `SessionProgressBar` to take an
`answerChoices` array and `numQuestions` variable instead of a single
"session data" object to make using the component more flexible.

Also, show the session progress bar in the session view. How we do that
(size, position, etc.) probably will change later and might be
configurable by the user.