Add trusted proxies config
Laravel's `url(...)` helper has the habit to return `http://...` URLs
and to disregard the configured `APP_URL` if the reverse proxy does use
http and if `X-Forwarded-Proto` is not set or trusted. Add
`config/trustedproxy.php` to allow users to trust proxies.

For some reason this approach is not documented in the official docs but
appears to be "the right way":

https://laravel.com/docs/10.x/requests#configuring-trusted-proxies
https://github.com/laravel/docs/pull/7457
https://github.com/laravel/framework/blob/10.x/src/Illuminate/Http/Middleware/TrustProxies.php

Resolves #556