SessionProgressBar: ignore answers with no matching question
An alternative would be to delete answer choices from sessions when the
deck changes, but this would mean that an edit to a public deck modifies
private user sessions, even if the edit was accidental or wrong and gets
reverted. So keep all answers instead and ignore the ones with no
matching question when calculating the session results.

Resolves #566