NewSessionView: refactor loading of deck data
Currently, we load all decks including their module and subjects on
mount. For installations with a lot of decks (e.g. 1K decks) this is
quite a bit of (unnecessary) data to load and slow.

Instead, only load the decks for the currently selected module. To do
that, add a new component `NewSessionDeckView` and change the controller
API endpoints accordingly.

This is mainly to resolve #583. The component structure and API design
remains a work in progress and we might revisit it later again.

Resolves #583