SessionView: don't error out on questions w/o answers
Usually we can expect questions to be well-formed (i.e. be of type
'card' or 'mc' and have 1 or 5 answers), but that's not always true for
"broken" questions imported from other systems. Make sure the answers
array isn't empty before accessing it, otherwise the frontend code
errors out and breaks the UI.