Add MagicGIF glitter
This is a silly but cute extra feature: when a user has answered 60% of
the questions correctly, they get to see a random GIF (think: positive
reinforcement). A similar feature exists in other (private) tools and
users like it.

Can be enabled/disabled globally via configuration as well as per user
in the settings.