DeckController: don't return `public-rw-listed` decks
Decks with `access` set to `public-rw-listed` are decks that "belong" to
all users of an OMC instance. They can be edited by opening them in the
editor from the deck view but shouldn't be in the list of "user owned"
decks.

Also, remove a `$request->wantsJson()` check; to get JSON, the API
controllers/endpoints should be used.