Add a "Repeat incorrect" flow for sessions
Add a `newFromIncorrect()` method to `Api/SessionController` which can
be used to create a new session with all questions that haven't been
correctly answered in the current session.

Also, rework the design of the session list on the start page to show an
open or continue button (depending on if the session has been completed)
as well as a "Repeat incorrect" button for sessions with incorrect
answers.

Design-wise we will probably refine that in the future.