NewSessionView: show statistics per module
Allow users to repeat the last 30/60/N incorrect questions of a module.

This is a first take and there's probably room for improvement.
Currently, the statistic is based on the latest sessions of the listed
decks. That could be confusing as users sometimes answer the questions
outside of the listed decks, for example in private decks or in the
"Repeat ..." sessions. Which answers should we consider here? To be
decided..