SessionCardAnswerView: hide answer buttons once answered
For questions of type "card", users can decide if they want to mark the
card as answered (or not) after they looked at the answer. Hide the
buttons once the question was marked. ("I got it" / "Ask me again")

Also, give the buttons a fixed width of 6 columns and unify the look.