NewSessionDeckView: ignore invalid questions in stats
So far invalid questions (questions marked as invalid) were counted into
the stats to make sure the number of questions shown in the stats
matches the sum of all questions of all decks. But this is confusing for
users, since they ignore invalid questions in their sessions and then
end up with unanswered questions in the stats.