Api/SessionController: handle sessions with deck changes
Currently, users encounter a client error and blank page if their
session points to a question which is no longer part of the underlying
deck. Check `session->current_question_id` on the controller side before
returning it to the user and set `current_question_id` to the first
question if the previous question has been deleted.