Add first stats to index page
Add a filled line graph with an "answers by hour" count for the last
seven days as well as a list of new and popular decks. There are a few
remaining TODOs and most likely we will adjust the style and shown data
later once we have a better understanding of what is of relevance and
useful.

Also, a few minor changes to unify the link style on the index page.