Models/User: fix `messages()` relationship method
`User->Message` is a one-to-many relationship, i.e. a user can have many
messages but a message can belong only to one user.

Also, since the foreign key column is named `author_id` and not
`user_id` (which would be the automatically determined name), we have to
explicitly specify the column in the `hasMany()` call. In the future, we
might rename the column instead to match the Eloquent default.

https://laravel.com/docs/10.x/eloquent-relationships#one-to-many