Add a first basic "Your comments" view
As we store dates and times in UTC, add a new config knob
`APP_TIMEZONE_VIEW` to make it possible to show formatted times in the
timezone of the instance. In the future we might change this to a user
configured timezone (in the frontend, we use date-fns which does format
according to user's timezone already) or use Svelte for the comments
view as well.

To "purify" HTML content in Blade templates, add htmlpurifier.