Session: only update current question when user opens editor
Currently, we update the question data in the background when the user
opens a question with the benefit that the shown data corresponds to the
latest version on the server at the time of opening.
The downsides are that we generate additional load and send many
requests for data that we already have (rate of edits is usually
significantly lower than number of users, sessions and answers) and that
we have to re-render the question view / form upon receiving the data
from the server.

Instead, use the question data as loaded at the beginning of the session
and only update the data if the user chooses to edit the question.

We have to see if and how this works out in the long run, but for the
moment it allows us to reduce load on the backend and at the same time
to implement "answer shuffling" (in a later commit) without delayed
re-rendering of the question due to data updates in the background.