Remove labels from doughnut charts
We have doughnut charts on both the "New session" page and "Session
Outro" component as a visual indicator of the results. As the labels
don't look very good and are not really needed to understand the chart,
remove them. Also, the design of both components will be reworked
soonish and all the info be given outside of the chart graph.