Make deck's `access` attribute editable and set default to 'private'
So far the `access` attribute was mostly a stub and not yet editable by
the user. Add a select element to the deck form to make it editable.
Also change the default `access` to private so that user collections are
private by default. Note that we don't plan to have private questions
currently, i.e. only the collection of questions is private but not the
contents of the individual questions in the deck.

Admins continue to be able to edit all decks, but we should add a
warning to decks where the admin user isn't the owner to avoid
accidental edits. (Will follow in a later commit.)

Also, add a few comments and extra lines to make the code more readable.