Implement soft deletion of comments
Currently, deletion of comments is a stub and not implemented yet. If
you click delete, we merely set the comment to an empty string. Instead,
add a flag `is_deleted` to the Message model and implement soft deletion
for comments.

As is this patch is an interim solution and we should put more thinking
and work into this later.

Using Laravel's `SoftDeletes` [1] maybe is a good idea, but we need to
keep in mind that we have to gently handle deletions of parent comments
(comments which have replies and therefore are important for the
conversation tree / flow). I.e. if we use `SoftDeletes`, we have to find
a different (placeholder?) approach for rendering of the comment tree
of a question.

[1] https://laravel.com/docs/11.x/eloquent#soft-deleting