QuestionView: fix `Uncaught TypeError`
Currently, when the user clicks the "Edit" button in the question view,
the editor doesn't open due to the following error:

```
Uncaught TypeError: updateCurrentQuestionData is not a function
```

With 5e1644cca862071daac570a65f20db06726f4aa3 we added a function
`updateCurrentQuestionData()` to update the data of the current question
when the user opens the editor. But in the question view context this
function isn't provided as there's no session context. Fix it by
checking if the function is available.

Resolves #784