IndexStatsView: fix `TypeError`
If OMC's cache was cleared or if its scheduled stats routines don't run,
the stats API doesn't return any data. If that's the case, the
IndexStatsView currently throws an error and the Svelte components on
the index page break:

```
TypeError: can't convert null to object
```

The issue can be reproduced by running `php artisan cache:clear` and
reloading the index page before new stats data was collected and put
into the cache.

Change the `IndexStatsView` component to check if data is available
before access to avoid the error.