IndexStatsView: fix "invalid date" error
The stats graph on the index page shows hourly answer and user counts on
hover. Though with Safari the label doesn't show the datetime but an
error message "invalid date". This is because the datetime string as
received from the API doesn't adhere to a standard datetime format (by
default, datetime fields are stored as UTC in the format
`2018-01-01 00:00:00`).

Update the scheduled stats routine to send datetime keys in ISO 8601
format (https://en.wikipedia.org/wiki/ISO_8601) and update the frontend
code to use date-fns's `parseISO` to parse the date from the label
context for the local time zone.

Also, change the index stats graph to show at most 4 ticks on the
y-axis.

Resolves #853