NewSessionView: use responsive offcanvas for subject/module selection
Currently, the subject and module selection is a `col-md-4` column. On
smartphones and screens with a small width, the selection is stacked on
top of the `col-md-8` for decks and users have to scroll past the list
of subjects and modules to get to the list of decks. Instead, use a
responsive offcanvas element for subject/module selection on small
screens which can be shown through a button.