resources/views/deck-editor: rename `Danger Zone` -> `Settings`
The settings currently listed under "Danger Zone" are not really
"dangerous" (no deletes happen), hence rename the section to "Settings"
to avoid confusion.

Also, remove the extra padding and styling to make the page look more
homogenous. We might update the style at a later point with more eye on
the overall design and its patterns.