SessionView: fix progress bar being shown in "exam mode"
With 05d3cb41e37292308ec4edc5ba4c6d4fbdd5539c we added a wrong if
statement whether the progress bar should be shown:

```
{#if !examMode && !sessionComplete || !currentQuestionContext.isAnswered}
```

Since the `&&` operator takes precedence over the `||` operator, the
statement will be interpreted as `(a && b) || c` and the progress bar
will be shown if the current question is not answered yet, no matter if
we are in exam mode (where the progress bar should be hidden throughout)
or not.

This should have been caught during review and testing but slipped
through.