SessionView: fix occasional `TypeError`
If a user instantly goes to the next question after selecting an answer
or if they are on a slow network connection, the
`currentQuestionContext` cannot be updated for the then previous
question after the answer was submitted. Fix it by optimistically
updating the question context first, before submitting the answer to the
server.

Resolves #899