Messages: fix message tree generation
Question comments get sorted and rendered into a "tree" (parent comments
and their replies). The current code doesn't work if a user replies to
an imported message (a message with `legacy_message_id`), as we use
`m.id` to reference and gather the childs but then look for childs for
`m.legacy_message_id`, i.e. we don't find the previously collected list
of childs. Use `m.id` throughout to fix the issue.