Add preliminary support for cases
Cases are useful if multiple questions share the same context (for
example clinical questions regarding a single patient in a medical
exam). This commit adds preliminary support: a data model, a controller
with the methods required to import data from another system and
frontend code to show the case text in the question view.

An update to the deck editor to add/modify cases will follow in a later
commit.