IndexStatsView: Add stats about last used decks
This adds a third column in the stats view to list the decks with the
most recent session creations.
I also removed the fixed width from the `col-md` classes. This way, the
other two (or one) column(s) can fill the screen if one (or two)
column(s) is missing, e.g. because of missing data or an error.