SessionQuestionView: don't shuffle again if user clicked "Show answer"
Currently, with answer shuffling enabled, if the user decides to use the
"Show answer" button to reveal the answer, the answers get shuffled a
second time. That's confusing and not intended.

The reason is that we update the questionContext which triggers a re-run
of the reactive statement where we shuffle answers. To fix it, update
the `if` statement that is supposed to make sure we only shuffle
unanswered questions upon load.