Kernel.php: Fix answer stats calculation
In the scheduler for the answer stats calculation, only answer choices
with related questions of the type MC should be retrieved by the
database query. In the existing code, all answer choices would be
queried, but only for those with MC type questions, the questions were
eager-loaded. This lead to wrong answer choices being included in the
collection and missing question objects.
This fixes this issue by actually only retrieving answer choices which
are associated to questions of the type MC.