Api/DeckController: make sure case is attached to deck
When a question with case gets added to a deck, we also have to attach
the case to the deck, since we expect related cases to be loaded and be
available in the deck editor.

Follow-up for 65e21e29796e271f288e2f1d7ed7cec2fe6a7fc0 ("Refactor deck
and question editor").