Revert "SessionQuestionView: sort answers by ID to ensure consistent order"
This reverts commit 4f0a98b15daaceed4f30c0a9e47ee2fabd537ce3.

Every selection of an answer triggers the reactive statements for
`question`: `question` is a bind to `currentQuestion` in the SessionView
and `currentQuestion` a reactive variable pointing at `data...` and
`data` is updated whenever an answer is given. So we can't do the answer
ordering in the reactive function as that does reorder answers on every
update of `data`. For the user this results in a unexpected reodering if
they have shuffling enabled.