Kernel.php: improve answer stats performance
Simplify the answer stats query and reduce the amount of retrieved data
from the database. Do as much data processing as possible in the
database. Specifically, only fetch a set of distinct questions for
processing to reduce the amount of data and therefore required memory in
the task worker.

We hope this avoids quickly growing, excessive memory usage and
eventually OOM errors during times when a lot of answer choices have to
be processed.