DeckController.php: Attach question only once
When using `->attach($question->id)`, it is possible to attach one
question to a deck multiple times because e.g. of missing UI feedback
even though there is not a valid use case for this.
Using `->syncWithoutDetaching($question->id)` solves this problem by
not re-attaching the question if it is already attached to the deck.