resources/views/decks: add sessions and access attribute as badges
Show the number of sessions as well as the `access` attribute on the
deck overview page.

Design wise this isn't optimal yet: the badges have the same color and
general look as the buttons above but are not clickable. Better button
and badge design and placement remains to be done later.