Api/QuestionMessageController: load `public_name` if set
After storing a new comment, we have to load the user's `public_name` to
make sure the frontend displays the public name instead of the username
(if a public name is set). Otherwise it looks like a bug to users and as
if the comment was stored with the real username instead of the public
name.

Resolves #1032