Refactor thumbs UI and data querying
Refactor the UI of the thumb up and down counter element in
`MessageView.svelte` to have less padding and be more clearly seperated
from other elements in the row.

Improve data retrieval in `QuestionMessageController.php` by counting
thumbs in the database and eager loading the author relationship for
all messages in one query.