DeckController: redirect to deck settings after creation
Currently, we reload the deck listing after creation of a new deck and
users don't get to see the deck settings. Some users might don't notice
that there are settings to control for example visibility of the deck.
Redirect users to the deck settings page in an attempt to make the
available options more known.