index: show loading spinners while loading data
Currently, we show a message "No sessions yet" or "No stats available
yet" while loading data. If loading the data takes a moment, for example
because the user is on a slow network, this can be confusing. Instead,
show a loading spinner while loading the data so that the user can
differentiate between "is still loading" and "no data".