Refactor and redesign comment section
Currently, the comment section looks rather crude and the indentation
and design doesn't work well on devices with a small screen width.

Refactor and improve the functionality and design as follows:

* Highlight / hide comments above / below a threshold of thumbs up /
  down
* Reduce the indentation and use border elements to display the comment
  thread
* Place the thumb up / down buttons and counter on the left side within
  the message element and reduce its size
* Reduce the size of the shown metadata (reply icon, author info, etc.)
* Use smaller buttons for Reply, Edit, etc.

Also, fix a rare bug where a "deleted comment" message would be shown
even when all replies are deleted as well.