github/workflows/ci: `git diff check` pull requests
Run `git diff --check <BASE>..HEAD` to "warn if changes introduce
conflict markers or whitespace errors".

This changes the triggers for our GitHub CI workflow in order to not run
jobs twice for pull requests, otherwise a push to a PR would trigger
both 'push' and 'pull_request'. Instead, CI jobs will now be triggered
on pull request events ('opened', 'synchronize', ...) and pushes to
master but not for other branches and events.