Upgrade to Svelte 5
Upgrade to Svelte 5 with its overhauled syntax and reactivity system:

https://svelte.dev/docs/svelte/v5-migration-guide

This patch is the sum of running

```
npx sv migrate svelte-5
```

on the `resources/` directory and the following manual updates and
adjustments:

* Remove invalid each item assignments:
  https://svelte.dev/docs/svelte/compiler-errors#each_item_invalid_assignment

* Remove a now forbidden `<p><div> ... </div></p>` construct (`div`
  inside `p` element)

* Add `$bindable()`  where the migration script didn't pick it up:
  https://svelte.dev/docs/svelte/v5-migration-guide#Breaking-changes-in-runes-mode-Bindings-need-to-be-explicitly-defined-using-$bindable()

* Use `untrack` (https://svelte.dev/docs/svelte/svelte#untrack) where
  necessary to retain existing behavior (e.g. `currentQuestionContext`
  updates)

For the time being, we use some of the helper functions provided by
`svelte/legacy`. Those will be migrated in a separate pull request (or
pull requests) over time.

Resolves #1011
Resolves #1102