Api/SessionController: abort if no questions found
In `newFromIncorrect()`, we only can create a new session if the current
session actually has incorrect answers. Fix the `if` statement to make
it abort if the collection is empty.

Otherwise we would encounter an error later on attempt to read the ID of
the first question (`$deck->questions()->first()->id`):

```
ErrorException: Attempt to read property "id" on null in .../app/Http/Controllers/Api/SessionController.php:78
```