index: reduce number of underscores and horizontal lines
With underscored headers and underscored deck links the index page shows
a lot of horizontal lines. Remove the underscores to make the page look
less cluttered.

Also, increase the vertical margin and reduce the vertical padding and
font size on the elements of the stats deck lists for a more compact and
clearer look.