Api/QuestionController: optimize answer choice updates
If the correct answer for a question was changed, we have to update all
existing answer choices for that question accordingly. Currently, we
read all previous answer choices and update them one-by-one, which is
both expensive (memory; number of DB queries) and unnecessary. Instead,
use a single SQL update query and let the database do the mass-update
efficiently.

There doesn't seem to be a way to do the update through Eloquent's API,
hence we use a raw `DB::statement`.

Manually tested on PostgreSQL, MariaDB and SQLite.

Fixes #575

Co-authored-by: Michael Schubert <schu@schu.io>