DeckQuestionController: Fix prev/next navigation
Decks and questions have a many-to-many relationship, the deck_question
pivot table defines the order of questions for a deck and therefore
questions (question IDs) are not strictly ascending from old to new
(from lower to higher ID) but mixed, in the order of their addition to
the deck.

This fixes the heuristic for finding the prev/next question in
DeckQuestionController by picking the prev/next question based on the
pivot table (i.e. order of addition to the deck) instead of using the
next lower/higher question ID.
In DeckController, the approach used to order questions is clarified.

Fixes #1203