composer update
Due to a change in the Laravel Framework with regards to route names and
precedence (presumably https://github.com/laravel/framework/pull/56920),
route precedence for the `users.edit` route changed and one of our
redirects and tests that relies on the old implicit route order fails.

```
return redirect()->route('users.edit', ...)
```

used to redirect to `users.edit` of our `UserManagementController` but
now redirects to our `Api\UserController`, i.e. the `Route::resource`
definition from `api.php` seems to be first now.

```
$ php artisan route:list -v | grep -A2 users.edit
  GET|HEAD        admin/users/{user}/edit .................................. users.edit › UserManagementController@edit
                  ⇂ web
                  ⇂ App\Http\Middleware\Authenticate:sanctum
--
  GET|HEAD        api/users/{user}/edit .......................................... users.edit › Api\UserController@edit
                  ⇂ api
                  ⇂ App\Http\Middleware\Authenticate:sanctum
```

As a fix, redirect to the controller action instead and don't rely on
the route name here:

https://laravel.com/docs/12.x/redirects#redirecting-controller-actions

As far as I can tell from the output below, the three fixed locations
are the only places where we have conflicting names in redirects.

```
$ php artisan route:list --json --sort name | jq -r .[].name | uniq -c | sort
[...]

$ git grep redirect app/Http/Controllers/
[...]
```

I wonder if this is the most idiomatic approach though and what others
with identical names for non-API and API routes do.