QuestionMessageController: Anonymize deleted comments
If a comment was deleted we still include info about the author via the
API endpoint. If a user deletes a comment, we should assume the user
wants it to be anonymized.
This changes the controller to not load (and therefore not provide it in
the response) the author info for deleted comments.