ApiQuestionMessageTest: adjust testDataIsAnonymized
assertJsonFragment accepts the key/value pair anywhere in the JSON,
in this case we want to be sure that the pair is located in the correct
message. Refactor to use AssertableJson to be explicit about this.
Include a deleted message in the test to check for anonymization there
as well.