SessionView: make sure `correct_answer_id` changes propagate immediately
Currently, if the users changes the correct answer for a question, this
change isn't immediately reflected in the UI and we continue to show the
old correct answer marked as correct and old progress bar state. To fix
this, add a Svelte `$effect` function to update the context and answer
choice for the current question whenever `currentQuestion` changes.

Using an `$effect` fuction to modify state is not lege artis
(https://svelte.dev/docs/svelte/$effect#When-not-to-use-$effect) but
seems to be acceptable here, as we don't modify question state (the
effect trigger) but the question context and answer choice, and both are
explicitly "untracked" (https://svelte.dev/docs/svelte/svelte#untrack)

We should try to refactor this as part of the work to remove
`svelte/legacy` code (see previous commit).