resources/views/info-single: display info posts in `alert-light` element
The info post tiles on the index and info overview page already have a
`alert-light` background. Put the contents of the info posts into a
`alert-light` element as well to give the content more structure and
make the look match the tiles. Also, make the info element `shadow-sm`,
similar to the contents of questions in the session view.