SessionQuestionView: add personal deck indicator
In the session view, indicate if the current question is already
included in a personal or bookmarked deck of the user by highlighting
the "AddToDeck" button with a distinct color (Bootstrap's `info` blue)
and also showing a label with the count of decks, if any (at the time of
loading the session).

A personal deck is a deck that is..

- owned by the current user,
- not public-rw-listed,
- not ephemeral and finally
- not archived.

A bookmarked deck is only shown if it's `public-rw`, since it doesn't
make sense to list decks that the current user cannot modify, i.e.
`public-ro` decks where they cannot add/remove questions.