Api/SessionController: fix performance regression in session loading
With e8315ebb36c5e63d8932936306085678fec4fe59 we added a "deck count
indicator" to the session view (if the question is part of 1 or more
personal decks, we highlight the "Add to deck" button and show a
counter), but the database query to aggregate the data is highly
inefficient due to the use of a single `OR` query, which prevents usage
of indexes and causes expensive full table scans. On big installations
with many questions and decks, the loading time is about 10x higher
(e.g. 250ms -> 2500ms).

Instead, run two separate queries without OR clauses, enabling index
usage on user_id and deck_id columns and merge the results in code.