AddToDeckDialog: allow limiting listed decks to current module only
Currently, users with many active decks in different modules have the
problem of a crammed "Add to deck" dialog. Add a config knob and toggle
to allow users to list only decks of the current module. Decks that have
the current question added are shown regardless of their module, to
avoid confusing users (e.g. question added to a deck according to "Add
to deck" label but no deck listed).

The filtering of the decks is done client-side, not on the server,
regardless of the user's preference we always get the full list of decks
from the server. This simplifies the implementation and allows us to
avoid repeated requests if the user toggles the config knob.

Also, change the URL path `decks/withquestionids` ->
`decks/with_question_ids` for better readability.

Resolves #1066